move scene viewer into separate mod
